% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit2.r
\name{digit2}
\alias{digit2}
\title{formating numbers}
\usage{
digit2(x, digit = 2, ...)
}
\arguments{
\item{x}{vector}

\item{digit}{a number defining digits}

\item{...}{...}
}
\description{
Return formated string of number
}
