% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff.r
\name{mean_diff}
\alias{mean_diff}
\title{mean difference output from t-tests}
\usage{
mean_diff(
  x,
  y,
  conf.level = 0.95,
  alternative = "two.sided",
  digit = 2,
  pdigits = 3,
  ...
)
}
\arguments{
\item{x}{group1}

\item{y}{group2}

\item{conf.level}{conf.level}

\item{alternative}{sided}

\item{digit}{digits for effect size}

\item{pdigits}{digits for p values}

\item{...}{...}
}
\description{
return ttest results: effect sizes and pvalues
}
