% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_ticks.r
\name{n_ticks}
\alias{n_ticks}
\title{ticks number control}
\usage{
n_ticks(n)
}
\arguments{
\item{n}{number of ticks}
}
\description{
ticks number control for ggplot
}
\examples{
"scale_x_continuous(break = n.ticks(20))"
}
