% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent2.R
\name{percent2}
\alias{percent2}
\title{formating numbers}
\usage{
percent2(x, digit = 2, ...)
}
\arguments{
\item{x}{vector}

\item{digit}{a number defining digits}

\item{...}{...}
}
\description{
Return formated string of number
}
