% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themec_yellow.R
\name{themec_yellow}
\alias{themec_yellow}
\title{ggplot2 theme yellow}
\usage{
themec_yellow(
  panel.aspect.ratio = 1,
  axis.title.textsize = 12,
  plot.tag.textsize = 14,
  plot.title.textsize = 14,
  plot.subtitle.textsize = 12,
  plot.footnote.textsize = 12,
  legend.position = "right",
  legend.textsize = 12,
  strip.textsize = 12
)
}
\arguments{
\item{panel.aspect.ratio}{panel.aspect.ratio}

\item{axis.title.textsize}{axis.title.textsize}

\item{plot.tag.textsize}{plot.tag.textsize}

\item{plot.title.textsize}{plot.title.textsize}

\item{plot.subtitle.textsize}{plot.subtitle.textsize}

\item{plot.footnote.textsize}{plot.footnote.textsize}

\item{legend.position}{legend.position = "top" or "right" or others}

\item{legend.textsize}{legend.textsize}

\item{strip.textsize}{strip.textsize}
}
\description{
clean and beautiful ggplot2 theme yellow old school
}
